Automation for the installation of the J2EE branch of the platform
OutSystemsAutomatedInstall.sh
This script allows a user with a CentOS6 box to quickly install the
requirements
for the platform 9.1. No inputs are needed.
No database engine is installed, we assume you have one ready for the
platform.

The script finishes by launching the configuration tool for user
customization of the database schemas, users and service ports.